Create Manage Logon Hour via Windows PowerShell
Hi,
First, thanks for reading this.
This post will discuss and teach how to set the user's logon hour via PowerShell.
I try to explain to a level of not over-explain the topic so the post won't be very long, and at the same time, showcase how things are going. I hope it will be clear.

There are some additional images, but I don't know how to upload the images to GitHub. The original post is on the following URL with the images
https://www.notion.so/invite/725e5391a14022f08f83fffc81435a20c1dba5f8
I use Notion.so for the markdown editing. 

Thanks again :)