(EDIT) Reflow Markdown
Prior to this change, the _Convert Specific Table of Excel Sheet to JSON_
post had lines that were occasionally very long. This edit reflows the
content, ensuring no line of Markdown is longer than 100 characters.

It uses the [Reflow Markdown][01] extension, which can reflow a block of
text to a specific line length.

Ensuring shorter lines makes it easier for editors to review the content,
to suggest edits, and to make diffs of changes easier for everyone to
review.

[01]: https://marketplace.visualstudio.com/items?itemName=marvhen.reflow-markdown