(EDIT) Format note & class name
This change updates the  _Convert Specific Table of Excel Sheet to JSON_
blog post for formatting. It replaces the paragraph that leads with
`Note -` with the [alert markup for a note][01].

It also corrects the identifier name, bolds it as a parameter, and reformats
the blog to always wrap `Class 6` in backticks as a specific value.

[01]: https://github.com/PowerShell/Community-Blog/blob/main/Docs/Markdown-cheatsheet.md#alert-boxes