(EDIT) Fix and standardize grammar.
This updates the _Convert Specific Table of Excel Sheet to JSON_ blog
post for minor grammar changes and clarity. A major change is to turn
the steps in the ordered list into sentences instead of pseudo-headings.

It also rewords the last heading from `Output` to the more semantically
accurate `Putting it all together`. This section shows the full code
and the output, not just the output. It acts as a conclusion to the
post.